Rename as middleware/reply-brain
