+++ app/routes/chatbot.js
@@ -12,7 +12,7 @@ const createUserMiddleware = require('../../lib/middleware/user-create');
 const createInboundMessageMiddleware = require('../../lib/middleware/message-inbound-create');
 const createOutboundMessageMiddleware = require('../../lib/middleware/message-outbound-create');
 const getBotReplyMiddleware = require('../../lib/middleware/bot-get-reply');
-const botReplyRivescriptMiddleware = require('../../lib/middleware/bot-reply-rivescript');
+const brainReplyMiddleware = require('../../lib/middleware/reply-brain');
 const botReplyMichaelMiddleware = require('../../lib/middleware/bot-reply-michael');
 const botReplySignupKeywordMiddleware = require('../../lib/middleware/bot-reply-signup-keyword');
 const botReplySignupMenuMiddleware = require('../../lib/middleware/bot-reply-signup-menu');
@@ -22,7 +22,7 @@ router.use(getUserMiddleware());
 router.use(createUserMiddleware());
 router.use(createInboundMessageMiddleware());
 router.use(getBotReplyMiddleware());
-router.use(botReplyRivescriptMiddleware());
+router.use(brainReplyMiddleware());
 router.use(botReplyMichaelMiddleware());
 router.use(botReplySignupKeywordMiddleware());
 router.use(botReplySignupMenuMiddleware());
