+++ lib/bot.js
@@ -60,7 +60,7 @@ module.exports.getBot = function getBot() {
  * @param {string} userMessage
  * @return {string}
  */
-module.exports.getRivescriptReply = function (user, userMessage) {
+module.exports.getReply = function (user, userMessage) {
   const bot = exports.getBot();
   bot.setUservars(user._id, { topic: user.topic });
