+++ lib/middleware/bot-get-reply.js
@@ -4,7 +4,7 @@ const bot = require('../bot.js');
 
 module.exports = function getReply() {
   return (req, res, next) => {
-    bot.getRivescriptReply(req.user, req.body.text)
+    bot.getReply(req.user, req.body.text)
       .then((text) => {
         req.reply = {
           rivescript: text,
